[RFC][WIP]Drivers: Intel: Updates to DMIC for runtime instantiate and remove DAI
This patch separates from function configure_registers() write
of FIFO A, FIFO B, and common functions initialization. The common
functions like CIC can be programmed for first instantiating. The
FIR A and the FIFO are programmed for FIFO A connected DAI and FIR B
related for FIFO B connected DAI.

The dmic_remove() is changed to pass the common resources stop and
free if there's still a running FIFO left.

The traced registers configuration are changed to 32 bit hex for easier
manual decode of bit fields when debugging.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>