[WIP][RFC] Drivers: DMIC: Fix a mistake in FIR length and update the set
The maximum length supported by HW is 250 and the existing decimate
by 5 set is 251. Load attempt of it failed always. However it was
needed in a rare 8 and 16 kHz FIFOs configuration request.

The new coefficients set fixes the issue. Also an alternate decimate
by 3 set is removed as unnecessary and a decimate by 10 set is added.
The iteration in FIR design was improved to be finer step so many of
the coefficients are changed despite of same specification.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>