[RFC][WIP] Drivers: DMIC: Make FIR decimation factor order configurable
This patch changes the logic in function select_mode() to use
the defined FIR decimation filters as order of preference for
FIR decimation factors. The earlier logic was simply to select the
lowest decimation factor. It can be used to e.g. prefer decimation
factor 6 over 5 when both are possible. The code behavior is not
yet changed because the filters are presented in increasing decimation
factor order.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>