topology: imx8: add demux component with wm8960 and cs42888 codecs
This topology works on i.MX8QM and i.MX8QXP with wm8960 and cs428888 codecs.
It adds a Multiplexer component for playback and a simple capture pipeline.

The MUX topology can be used for test bench and real device usage.

I've also added a pipeline for DAI playback starting with DEMUX.

Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>